pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "FACE01"
version = "2.2.02"
requires-python = ">=3.8"
dependencies = [
    "GPUtil",
    "Pillow",
    "dlib==19.24.0",
    "faiss-gpu",
    "mediapipe",
    "memory-profiler",
    "mojimoji",
    "nptyping",
    "numpy",
    "onnx",
    "onnxruntime-gpu",
    "opencv-python",
    "protobuf",
    "psutil",
    "pyqrcode",
    "qrcode",
    "requests",
    "scipy",
    "snakeviz",
    "torch",
    "torchvision",
    "tqdm",
    "ttkbootstrap",
    "typing_extensions",
    "urllib3",
    "watchdog"
]
license = {file = "LICENSE"}
readme = "README.md"
description = "顔認証学習モデルとユーティリティプログラム"
authors = [
    { name="Yoshitsugu Kesamaru", email="y.kesamaru@tokai-kaoninsho.com" }
]
keywords = ["face recognition"]

[project.urls]
Homepage = "https://tokai-kaoninsho.com/"
Documentation = "https://ykesamaru.github.io/FACE01_SAMPLE/"
Repository = "https://github.com/yKesamaru/FACE01_SAMPLE"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["face01lib"]

